+++ CHANGELOG.rst
@@ -20,9 +20,14 @@ Added
 * Add new ``?tags``, query param filter to the ``/v1/actions`` API endpoint. This query parameter
   allows users to filter out actions based on the tag name . By default, when no filter values are
   provided, all actions are returned. (new feature) #4219
-* Allow user to force terminal size to use by ``st2`` CLI for table formatting purposes by setting
-  ``ST2_CLI_FORCE_TERMINAL_SIZE=<lines,column>`` environment variable. For example,
-  ``ST2_CLI_FORCE_TERMINAL_SIZE=80,200 st2 action list``. (improvement) #4242
+* Update ``st2`` CLI to it inspect ``LINES`` and ``COLUMNS`` environment variables first when
+  determining the terminal size. Previously those environment variables were checked second last
+  (after trying to retrieve terminal size using various OS specific methods and before falling back
+  to the default values).
+
+  This approach is more performant and allows user to easily overwrite the default values or values
+  returned by the operating system checks - e.g. by running
+  ``LINES=80 COLUMNS=200 st2 action list``. (improvement) #4242
 
 Changed
 ~~~~~~~